% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{almonds_bowl}
\alias{almonds_bowl}
\title{Chocolate-covered almonds data}
\format{
A data frame with 5000 observations on the following 2 variables
\describe{
\item{ID}{Identification value for a given chocolate-covered almond}
\item{weight}{Weight of the chocolate-covered almond in grams (to the nearest tenth)}
}
}
\usage{
almonds_bowl
}
\description{
5000 chocolate-covered almonds selected from a large batch, weighed in grams.
}
\keyword{datasets}
