% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{almonds_sample}
\alias{almonds_sample}
\title{Chocolate-covered almonds data sample of size 25}
\format{
A data frame with 25 observations on the following 2 variables
\describe{
\item{replicate}{Replicate number set to 1 since there is only one sample}
\item{ID}{Identification value for a given chocolate-covered almond}
\item{weight}{Weight of the chocolate-covered almond in grams (to the nearest tenth)}
}
}
\usage{
almonds_sample
}
\description{
A sample of 25 chocolate-covered almonds, weighed in grams.
}
\keyword{datasets}
