% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{amazon_books}
\alias{amazon_books}
\title{Sample of Amazon books}
\format{
A data frame of 325 rows representing books listed on Amazon and 13 variables.
\describe{
\item{title}{Book title}
\item{author}{Author who wrote book}
\item{list_price}{recommended retail price of book}
\item{amazon_price}{lowest price of book shown on Amazon}
\item{hard_paper}{book is either hardcover or paperback}
\item{num_pages}{number of pages in book}
\item{publisher}{Company that issues the book for sale}
\item{pub_year}{Year the book was published}
\item{isbn_10}{10-character ISBN number}
\item{height, width, thick, weight_oz}{height, width, weight and thickness of the book}
}
}
\source{
The Data and Story Library (DASL) \url{https://dasl.datadescription.com/datafile/amazon-books}
}
\usage{
amazon_books
}
\description{
A random sample of 325 books from Amazon.com.
}
\keyword{datasets}
