% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{avocados}
\alias{avocados}
\title{Avocado Prices by US Region}
\format{
A data frame of 54 regions over 3 years of weekly results
\describe{
\item{date}{Week of Data Recording}
\item{average_price}{Average Price of Avocado}
\item{total_volume}{Total Amount of Avocados}
\item{small_hass_sold}{Amount of Small Haas Avocados Sold}
\item{large_hass_sold}{Amount of Large Haas Avocados Sold}
\item{xlarage_hass_sold}{Amount of Extra Large Haas Avocados Sold}
\item{total_bags}{Total Amount of Bags of Avocados}
\item{small_bags}{Total Amount of Bags of Small Haas Avocados}
\item{large_bags}{Total Amount of Bags of Large Haas Avocados}
\item{x_large_bags}{Total Amount of Bags of Extra Large Haas Avocados}
\item{type}{Type of Sale}
\item{year}{Year of Sale}
\item{region}{Region Where Sale Took Place}
}
}
\usage{
avocados
}
\description{
Gathered from \url{https://docs.google.com/spreadsheets/d/1cNuj9V-9Xe8fqV3DQRhvsXJhER3zTkO1dSsQ1Q0j96g/edit#gid=1419070688}
}
\keyword{datasets}
