% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ball_samples.R
\docType{data}
\name{ball_samples}
\alias{ball_samples}
\title{Sampling from a tub of balls}
\format{A data frame 10 rows representing different groups of students' 
samples of size n = 50 and 5 variables
\describe{
  \item{group}{Group name}
  \item{red}{Number of red balls sampled}
  \item{white}{Number of white balls sampled}
  \item{green}{Number of green balls sampled}
  \item{n}{Total number of balls samples}
}}
\usage{
ball_samples
}
\description{
Counting the number of red balls in 10 samples of size n = 50 balls from
\url{https://github.com/ismayc/moderndiver-book/raw/master/images/sampling1.jpg}
}
\examples{
library(dplyr)
library(ggplot2)

# Compute proportion red
red_ball_samples <- red_ball_samples \%>\%
  mutate(prop_red = red / n)
  
# Plot sampling distributions
ggplot(red_ball_samples, aes(x = prop_red)) +
  geom_histogram(binwidth = 0.05) +
  labs(x = expression(hat(p)), y = "Number of samples", 
  title = "Sampling distribution of p_hat based 10 samples of size n = 50")
}
\keyword{datasets}
