% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bowl}
\alias{bowl}
\title{A sampling bowl of red and white balls}
\format{
A data frame 2400 rows representing different balls in the bowl, of which
900 are red and 1500 are white.
\describe{
  \item{ball_ID}{ID variable used to denote all balls. Note this value is not
  marked on the balls themselves}
  \item{color}{color of ball: red or white}
}
}
\usage{
bowl
}
\description{
A sampling bowl used as the population in a simulated sampling exercise. Also
known as the urn sampling framework \url{https://en.wikipedia.org/wiki/Urn_problem}.
}
\keyword{datasets}
