% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{get_regression_points}
\alias{get_regression_points}
\title{Get regression points}
\usage{
get_regression_points(model, digits = 3, print = FALSE, newdata = NULL)
}
\arguments{
\item{model}{a model object (Currently only \code{lm} is supported)}

\item{digits}{number of digits precision in output table}

\item{print}{If TRUE, return in print format suitable for R Markdown}

\item{newdata}{a new data frame to fit the regression to/make predictions on.}
}
\value{
A tibble or nicely formatted table
}
\description{
Wrapper for tidy \code{lm()} regression function fit output
}
\examples{
library(dplyr)
library(tibble)

# Get information on all points in regression
mpg_model <- lm(mpg ~ cyl, data = mtcars)
get_regression_points(mpg_model)

# Convert rownames to column
data("mtcars")
mtcars <- mtcars \%>\% 
  rownames_to_column(var = "model")

# Create training and test set.
training_set <- mtcars \%>\% 
  sample_frac(0.5)
test_set <- mtcars \%>\% 
  anti_join(training_set, by = "model")

# Fit model to training set, predict on test set
mpg_model_train <- lm(mpg ~ cyl, data = training_set)
get_regression_points(mpg_model_train, newdata = test_set)
}
