% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ipf_lifts}
\alias{ipf_lifts}
\title{International Power Lifting Results}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 41152 rows and 16 columns.
}
\usage{
ipf_lifts
}
\description{
International Power Lifting Results
}
\keyword{datasets}
