% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mass_traffic_2020}
\alias{mass_traffic_2020}
\title{2020 road traffic volume and crash level date for 13 Massachusetts counties}
\format{
A data frame of 874 rows representing traffic data at the 874 sites
\describe{
\item{site_id}{Site id}
\item{county}{County in which the site is located}
\item{community}{Community in which the site is located}
\item{rural_urban}{Rural (R) or Urban (U)}
\item{dir}{Direction for traffic movement. Either 1-WAY, 2-WAY, EB (eastbound), RAMP or WB (westbound)}
\item{functional_class}{Classification of road. Either Arterial, Collector, Freeway & Expressway, Interstate or Local Road}
\item{avg_speed}{Average traffic speed}
\item{total_volume}{Number of vehicles recorded at each site in 2020}
\item{crashes}{Number of vehicle crashes at each site}
\item{nonfatal_injuries}{Number of non-fatal injuries for all recorded vehicle crashes}
\item{fatal_injuries}{Number of fatal injuries for all recorded vehicle crashes}
}
}
\usage{
mass_traffic_2020
}
\description{
2020 road traffic volume and crash level date for 13 Massachusetts counties
}
\keyword{datasets}
