% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{movies_sample}
\alias{movies_sample}
\title{Random sample of 68 action and romance movies}
\format{
A data frame of 68 rows movies.
\describe{
  \item{title}{Movie title}
  \item{year}{Year released}
  \item{rating}{IMDb rating out of 10 stars}
  \item{genre}{Action or Romance}
}
}
\usage{
movies_sample
}
\description{
A random sample of 32 action movies and 36 romance movies from
\url{https://www.imdb.com/} and their ratings.
}
\examples{
library(ggplot2)

# Visualize relationship between rating and genre
ggplot(data = movies_sample, aes(x = genre, y = rating)) +
  geom_boxplot() +
  labs(x = "Genre: Action or Romance", y = "IMDb rating")
}
\seealso{
This data was sampled from the `movies` data frame in the \code{ggplot2movies} package.
}
\keyword{datasets}
