% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mythbusters_yawn}
\alias{mythbusters_yawn}
\title{Data from Mythbusters' study on contagiousness of yawning}
\format{
A data frame of 50 rows representing each of the 50 participants
in the study.
\describe{
  \item{subj}{integer value corresponding to identifier variable of 
  subject ID}
  \item{group}{string of either \code{"seed"}, participant was shown a 
  yawner, or \code{"control"}, participant was not shown a yawner}
  \item{yawn}{string of either \code{"yes"}, the participant yawned, or
  \code{"no"}, the participant did not yawn}
}
}
\usage{
mythbusters_yawn
}
\description{
From a study on whether yawning is contagious
\url{https://www.imdb.com/title/tt0768479/}.
The data here was derived from the final proportions of yawns given
in the show.
}
\examples{
library(ggplot2)

# Plot both variables as a stacked proportional bar chart
ggplot(mythbusters_yawn, aes(x = group, fill = yawn)) +
  geom_bar(position = "fill") +
  labs(x = "", y = "Proportion", 
  title = "Proportion of yawn and not yawn for each group")
}
\keyword{datasets}
