% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{old_faithful_2024}
\alias{old_faithful_2024}
\title{Old Faithful Eruptions Dataset (2024)}
\format{
A data frame with 114 rows and 6 variables:
\describe{
\item{eruption_id}{\code{numeric}. A unique identifier for each eruption.}
\item{date}{\code{date}. The date of the eruption.}
\item{time}{\code{numeric}. The time of the eruption in HHMM format (e.g., 538 corresponds to 5:38 AM, 1541 corresponds to 3:41 PM).}
\item{waiting}{\code{numeric}. The waiting time in minutes until the next eruption.}
\item{webcam}{\code{character}. Indicates whether the eruption was captured on webcam ("Yes" or "No").}
\item{duration}{\code{numeric}. The duration of the eruption in seconds.}
}
}
\source{
Volunteer information from https://geysertimes.org/retrieve.php
}
\usage{
old_faithful_2024
}
\description{
This dataset contains records of eruptions from the Old Faithful geyser in Yellowstone National Park, recorded in 2024.
It includes details such as the eruption ID, date and time of eruption, waiting time between eruptions, webcam availability,
and the duration of each eruption.
}
\keyword{datasets}
