% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pennies}
\alias{pennies}
\title{A population of 800 pennies sampled in 2011}
\format{
A data frame of 800 rows representing different pennies and 2 variables
\describe{
  \item{year}{Year of minting}
  \item{age_in_2011}{Age in 2011}
}
}
\source{
StatCrunch \url{https://www.statcrunch.com/app/index.php?dataid=301596}
}
\usage{
pennies
}
\description{
A dataset of 800 pennies to be treated as a sampling population. Data on
these pennies were recorded in 2011.
}
\keyword{datasets}
