% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pennies_resamples}
\alias{pennies_resamples}
\title{Bootstrap resamples of a sample of 50 pennies}
\format{
A data frame of 1750 rows representing 35 students' bootstrap
resamples of size 50 and 3 variables
\describe{
\item{replicate}{ID variable of replicate/resample number.}
\item{name}{Name of student}
\item{year}{Year on resampled penny}
}
}
\usage{
pennies_resamples
}
\description{
35 bootstrap resamples with replacement of sample of 50 pennies contained in
a 50 cent roll from Florence Bank on Friday February 1, 2019 in downtown Northampton,
Massachusetts, USA \url{https://www.google.com/maps/place/Florence+Bank+-+Downtown+Northampton/@42.3194512,-72.6337622,16z/data=!4m5!3m4!1s0x89e6d7465ea57a19:0xc901966a46d8ceab!8m2!3d42.3194512!4d-72.6293848}. The original sample
of 50 pennies is available in \code{\link[=pennies_sample]{pennies_sample()}} .
}
\seealso{
\code{\link[=pennies_sample]{pennies_sample()}}
}
\keyword{datasets}
