% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pennies_sample}
\alias{pennies_sample}
\title{A sample of 50 pennies}
\format{
A data frame of 50 rows representing 50 sampled pennies and 2 variables
\describe{
\item{ID}{Variable used to uniquely identify each penny.}
\item{year}{Year of minting.}
}
}
\usage{
pennies_sample
}
\description{
A sample of 50 pennies contained in a 50 cent roll from Florence Bank on
Friday February 1, 2019 in downtown Northampton, Massachusetts, USA
\url{https://www.google.com/maps/place/Florence+Bank+-+Downtown+Northampton/@42.3194512,-72.6337622,16z/data=!4m5!3m4!1s0x89e6d7465ea57a19:0xc901966a46d8ceab!8m2!3d42.3194512!4d-72.6293848}.
}
\note{
The original \code{pennies_sample} has been renamed \code{\link[=orig_pennies_sample]{orig_pennies_sample()}}
as of \code{moderndive} v0.3.0.
}
\keyword{datasets}
