% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{saratoga_houses}
\alias{saratoga_houses}
\title{House Prices and Properties in Saratoga, New York}
\format{
A data frame with 1057 observations on the following 8 variables
\describe{
\item{price}{price (US dollars)}
\item{living_area}{Living Area (square feet)}
\item{bathrooms}{Number of Bathroom (half bathrooms have no shower or tub)}
\item{bedrooms}{Number of Bedrooms}
\item{fireplaces}{Number of Fireplaces}
\item{lot_size}{Size of Lot (acres)}
\item{age}{Age of House (years)}
\item{fireplace}{Whether the house has a Fireplace}
}
}
\source{
Data on 2006 house prices in Saratoga County, New York,
USA. Originally gathered from a publicly shared Google Sheet that is
no longer publicly accessible.
}
\usage{
saratoga_houses
}
\description{
Random sample of 1057 houses taken from full Saratoga Housing Data (De Veaux)
}
\keyword{datasets}
