% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{almonds_sample_100}
\alias{almonds_sample_100}
\title{Chocolate-covered almonds data sample}
\format{
A data frame with 100 observations on the following 2 variables
\describe{
\item{ID}{Identification value for a given chocolate-covered almond}
\item{weight}{Weight of the chocolate-covered almond in grams (to the nearest tenth)}
}
}
\usage{
almonds_sample_100
}
\description{
A sample of 100 chocolate-covered almonds, weighed in grams.
}
\keyword{datasets}
