% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowl.R
\docType{data}
\name{bowl_sample_1}
\alias{bowl_sample_1}
\title{Tactile sample of size 50 from a bowl of balls}
\format{
A data frame of 50 rows representing different balls and 1 variable.
\describe{
  \item{color}{Color of ball sampled}
}
}
\usage{
bowl_sample_1
}
\description{
A single tactile sample of size n = 50 balls from
\url{https://github.com/moderndive/moderndive/blob/master/data-raw/sampling_bowl.jpeg}
}
\examples{
library(ggplot2)

# Barplot of distribution of balls in sample
ggplot(bowl_sample_1, aes(x = color)) +
  geom_bar() +
  labs(title = "50 sampled bals from bowl")
}
\seealso{
\code{\link{bowl}}
}
\keyword{datasets}
