% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowl.R
\docType{data}
\name{bowl_samples}
\alias{bowl_samples}
\title{Sampling from a bowl of balls}
\format{
A data frame 10 rows representing different groups of students'
samples of size n = 50 and 5 variables
\describe{
  \item{group}{Group name}
  \item{red}{Number of red balls sampled}
  \item{white}{Number of white balls sampled}
  \item{green}{Number of green balls sampled}
  \item{n}{Total number of balls samples}
}
}
\usage{
bowl_samples
}
\description{
Counting the number of red balls in 10 samples of size n = 50 balls from
\url{https://github.com/moderndive/moderndive/blob/master/data-raw/sampling_bowl.jpeg}
}
\examples{
library(dplyr)
library(ggplot2)

# Compute proportion red
bowl_samples <- bowl_samples \%>\%
  mutate(prop_red = red / n)

# Plot sampling distributions
ggplot(bowl_samples, aes(x = prop_red)) +
  geom_histogram(binwidth = 0.05) +
  labs(
    x = expression(hat(p)), y = "Number of samples",
    title = "Sampling distribution of p_hat based 10 samples of size n = 50"
  )
}
\seealso{
\code{\link{bowl}}
}
\keyword{datasets}
