% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{early_january_weather}
\alias{early_january_weather}
\title{Early January hourly weather data}
\format{
A data frame of 358 rows representing hourly measurements and 15 variables
\describe{
\item{origin}{Weather station. Named \code{origin} to facilitate merging with
\code{\link[nycflights13:flights]{nycflights13::flights}} data.}
\item{year, month, day, hour}{Time of recording.}
\item{temp, dewp}{Temperature and dewpoint in F.}
\item{humid}{Relative humidity.}
\item{wind_dir, wind_speed, wind_gust}{Wind direction (in degrees), speed
and gust speed (in mph).}
\item{precip}{Precipitation, in inches.}
\item{pressure}{Sea level pressure in millibars.}
\item{visib}{Visibility in miles.}
\item{time_hour}{Date and hour of the recording as a \code{POSIXct} date.}
}
}
\source{
ASOS download from Iowa Environmental Mesonet,
\url{https://mesonet.agron.iastate.edu/request/download.phtml}.
}
\usage{
early_january_weather
}
\description{
Hourly meterological data for LGA, JFK and EWR for the month of January 2013.
This is a subset of the \code{weather} data frame from \code{nycflights13}.
}
\seealso{
\code{\link[nycflights13:weather]{nycflights13::weather}}.
}
\keyword{datasets}
