% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_parallel_slopes.R
\name{geom_parallel_slopes}
\alias{geom_parallel_slopes}
\title{Parallel slopes model}
\usage{
geom_parallel_slopes(mapping = NULL, data = NULL,
  position = "identity", ..., se = TRUE, formula = y ~ x, n = 100,
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{se}{Display confidence interval around model lines? \code{TRUE} by
default.}

\item{formula}{Formula to use per group in parallel slopes model. Basic
linear \code{y ~ x} by default.}

\item{n}{Number of points per group at which to evaluate model.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}
}
\description{
\code{geom_parallel_slopes()} fits parallel slopes model and adds its line
output(s) to a \code{ggplot} object. Basically, it fits a unified model with
intercepts varying between groups (which should be supplied as standard
{ggplot2} grouping aesthetics: \code{group}, \code{color}, \code{fill},
etc.). This function has the same nature as \code{geom_smooth()} from
{ggplot2} package, but provides functionality that \code{geom_smooth()}
currently doesn't have.
}
\examples{
library(dplyr)
library(ggplot2)

example_df <- house_prices \%>\%
  slice(1:1000) \%>\% 
  mutate(
    log10_price = log10(price),
    log10_size = log10(sqft_living)
  )
ggplot_viz <- ggplot(
  data = example_df,
  mapping = aes(x = log10_size, y = log10_price, color = condition)
) +
  geom_point(alpha = 0.1)

# Basic usage
ggplot_viz + geom_parallel_slopes()
ggplot_viz + geom_parallel_slopes(se = FALSE)

# Supply custom aesthetics
ggplot_viz + geom_parallel_slopes(size = 4)

# Fit non-linear model
ggplot_viz + geom_parallel_slopes(formula = y ~ poly(x, 2))

# Different grouping
ggplot(example_df, aes(x = log10_size, y = log10_price)) +
  geom_point(alpha = 0.1) +
  geom_parallel_slopes(aes(fill = condition))

}
