% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{get_regression_points}
\alias{get_regression_points}
\title{Get regression points}
\usage{
get_regression_points(formula, data, digits = 3, print = FALSE, ...)
}
\arguments{
\item{formula}{regression formula}

\item{data}{data frame}

\item{digits}{number of digits precision in output table}

\item{print}{If true, return in print format suitable for R Markdown}

\item{...}{other arguments passed to lm()}
}
\value{
table
}
\description{
Wrapper for \code{lm()} regression function
}
\examples{
get_regression_points(mpg ~ cyl, data = mtcars)
}
