% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_sample_n.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{rep_sample_n}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{infer}{\code{\link[infer]{rep_sample_n}}}
}}

